* change category field of oa_relation.
git-svn-id: file:///tmp/ranzhi/trunk@989 cb094fd3-5a31-0410-bd9d-ef5a9cbf4261

+++ ranzhi/db/ranzhi.sql
@@ -252,7 +252,7 @@ CREATE TABLE `oa_docLib` (
 CREATE TABLE `oa_relation` (
   `type` char(20) NOT NULL,
   `id` mediumint(9) NOT NULL,
-  `category` smallint(5) NOT NULL,
+  `category` mediumint(8) NOT NULL,
   UNIQUE KEY `relation` (`type`,`id`,`category`)
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
 -- DROP TABLE IF EXISTS `sys_action`;
